+++ bindings/nodejs/examples/how_tos/native_tokens/mint.ts
@@ -1,22 +1,18 @@
 // Copyright 2023 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-import { MintNativeTokenParams, utf8ToHex } from '@iota/sdk';
-
 import { getUnlockedWallet } from '../../wallet/common';
 
-// The circulating supply of the native token. `100` hex encoded
-const CIRCULATING_SUPPLY = '0x64';
-// The maximum supply of the native token. `100` hex encoded
-const MAXIMUM_SUPPLY = '0x64';
+// The amount of native tokens to mint, 10 hex encoded.
+const MINT_AMOUNT = '0xA';
 
-// In this example we will mint a native token.
+// In this example we will mint an existing native token with its foundry.
 //
 // Make sure that `example.stronghold` and `example.walletdb` already exist by
 // running the `how_tos/accounts-and-addresses/create-wallet` example!
 //
 // Rename `.env.example` to `.env` first, then run
-// yarn run-example ./wallet/09-mint-native-token.ts
+// yarn run-example ./how_tos/native_tokens/mint.ts
 async function run() {
     try {
         // Create the wallet
@@ -25,40 +21,31 @@ async function run() {
         // Get the account we generated with `01-create-wallet`
         const account = await wallet.getAccount('Alice');
 
-        const balance = await account.sync();
+        // May want to ensure the account is synced before sending a transaction.
+        let balance = await account.sync();
 
-        // We can first check if we already have an alias in our account, because an alias can have many foundry outputs and therefore we can reuse an existing one
-        if (balance.aliases.length > 0) {
-            // If we don't have an alias, we need to create one
-            const transaction = await account
-                .prepareCreateAliasOutput()
-                .then((prepared) => prepared.send());
-            console.log(`Transaction sent: ${transaction.transactionId}`);
-
-            // Wait for transaction to get included
-            const blockId = await account.retryTransactionUntilIncluded(
-                transaction.transactionId,
-            );
-
-            console.log(
-                `Block included: ${process.env.EXPLORER_URL}/block/${blockId}`,
-            );
-
-            await account.sync();
-            console.log('Account synced');
+        if (balance.foundries.length == 0) {
+            throw new Error(`No Foundry available in account 'Alice'`);
         }
 
-        console.log('Preparing minting transaction...');
+        // Find first foundry and corresponding token id
+        const tokenId = balance.foundries[0];
 
-        // If we omit the AccountAddress field the first address of the account is used by default
-        const params: MintNativeTokenParams = {
-            circulatingSupply: CIRCULATING_SUPPLY,
-            maximumSupply: MAXIMUM_SUPPLY,
-            foundryMetadata: utf8ToHex('Hello, World!'),
-        };
+        let token = balance.nativeTokens.find(
+            (nativeToken) => nativeToken.tokenId == tokenId,
+        );
+        if (token == null) {
+            throw new Error(
+                `Couldn't find native token '${tokenId}' in the account`,
+            );
+        }
 
-        const prepared = await account.prepareMintNativeToken(params);
-        const transaction = await prepared.send();
+        console.log(`Balance before minting:`, parseInt(token.available));
+
+        // Mint some more native tokens
+        const transaction = await account
+            .prepareMintNativeToken(token.tokenId, MINT_AMOUNT)
+            .then((prepared) => prepared.send());
 
         console.log(`Transaction sent: ${transaction.transactionId}`);
 
@@ -71,11 +58,16 @@ async function run() {
             `Block included: ${process.env.EXPLORER_URL}/block/${blockId}`,
         );
 
-        console.log(`Minted token: ${prepared.tokenId()}`);
-
-        // Ensure the account is synced after minting.
-        await account.sync();
-        console.log('Account synced');
+        balance = await account.sync();
+        token = balance.nativeTokens.find(
+            (nativeToken) => nativeToken.tokenId == tokenId,
+        );
+        if (token == null) {
+            throw new Error(
+                `Couldn't find native token '${tokenId}' in the account`,
+            );
+        }
+        console.log(`Balance after minting:`, parseInt(token.available));
     } catch (error) {
         console.log('Error: ', error);
     }