+++ bindings/python/iota_sdk/client/_high_level_api.py
@@ -1,9 +1,44 @@
 # Copyright 2023 IOTA Stiftung
 # SPDX-License-Identifier: Apache-2.0
 
+from iota_sdk.secret_manager.secret_manager import LedgerNanoSecretManager, MnemonicSecretManager, StrongholdSecretManager, SeedSecretManager
 from iota_sdk.types.common import HexStr
 from iota_sdk.types.output_id import OutputId
+from iota_sdk.types.common import CoinType
 from typing import List, Optional
+
+
+class Range:
+    def __init__(self, start: int, end: int):
+        self.start = start
+        self.end = end
+
+
+class GenerateAddressOptions():
+    def __init__(self,  internal: bool, ledgerNanoPrompt: bool):
+        self.internal = internal
+        self.ledgerNanoPrompt = ledgerNanoPrompt
+
+
+class GenerateAddressesOptions():
+    def __init__(self, coinType: CoinType,
+                 range: range,
+                 bech32Hrp: str,
+                 accountIndex: Optional[int] = None,
+                 options: Optional[GenerateAddressOptions] = None):
+        self.coinType = coinType
+        self.range = Range(range.start, range.stop)
+        self.bech32Hrp = bech32Hrp
+        self.accountIndex = accountIndex
+        self.options = options
+
+    def as_dict(self):
+        config = {k: v for k, v in self.__dict__.items() if v != None}
+
+        config["range"] = config["range"].__dict__
+        if "options" in config:
+            config["options"] = config["options"].__dict__
+        return config
 
 
 class HighLevelAPI():
@@ -47,13 +82,13 @@ class HighLevelAPI():
             'maxAttempts': max_attempts
         })
 
-    def consolidate_funds(self, secret_manager, generate_addresses_options):
+    def consolidate_funds(self, secret_manager: LedgerNanoSecretManager | MnemonicSecretManager | SeedSecretManager | StrongholdSecretManager, generate_addresses_options: GenerateAddressesOptions):
         """Function to consolidate all funds from a range of addresses to the address with the lowest index in that range
            Returns the address to which the funds got consolidated, if any were available.
         """
         return self._call_method('consolidateFunds', {
             'secretManager': secret_manager,
-            'generateAddressesOptions': generate_addresses_options,
+            'generateAddressesOptions': generate_addresses_options.as_dict(),
         })
 
     def find_inputs(self, addresses: List[str], amount: int):