+++ bindings/nodejs/CHANGELOG.md
@@ -19,6 +19,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## 1.0.0-rc.2 - 2023-0x-xx
+
+### Changed
+
+- Rename `Account::prepareMintNativeToken` to `prepareCreateNativeToken`, `Account::prepareIncreaseNativeTokenSupply` to `prepareMintNativeToken`, `Account::prepareDecreaseNativeTokenSupply` to `prepareMeltNativeToken`;
+- Rename `MintNativeTokenParams` to `CreateNativeTokenParams`;
+- Rename `MintTokenTransaction` to `CreateNativeTokenTransaction` and `PreparedMintTokenTransaction` to `PreparedCreateNativeTokenTransaction` (including their corresponding `Data` types);
+
+### Fixed
+
+- Moved `internal` field from `IGenerateAddressesOptions` to `IGenerateAddressOptions`;
+- Error handling in `Client`, `SecretManager` and `Wallet` constructors;
+
 ## 1.0.0-rc.1 - 2023-06-19
 
 ### Added